[SWM-341] Fix : Text Loop
